requirements.txt
python-whois
requests
argparse
